requirements.txt
google-auth
google-cloud-core
google-cloud-storage
google-cloud-storage-control
psycopg